LIB/requirements.txt
robotframework
robotframework-httplibrary
urllib3
requests
jsonschema